Added initial script for verifying the algorithm
The script sets up a predefined triangle rotated 55 degree counterclockwise.
It then simulates a user dragging one of the rectangles corners to a new location -2,2 away from the corners previous position. 
It then executes the algorithm in order to calculate the new rectangle's shape and draws the steps in a plot.